feature:#LFY-10 Fix for CompleteVerison Parsing.
CompleteVersion parsing does not respect the different digits by default, modified constructor to allow for different prefixes just to simplify unit testing cases being needed for LFY-10.  Note it does not correcty set the prefix but it can parse it now.  Other refactorings and housekeeping included including adding the logo style output and version to make it easier to see which version we are using.